Rough compatibility update to fix the stuff iOS 10 damaged
- FIXED: Navigation BarButtonItems are now available/visible again
- FIXED: Issues with Alerts and Actionsheets
- FIXED: Countless warnings for some manic type precision voodoo of Xcode
- ADDED: Some better icon for settings & back nav

+++ Colors/UIColor+Expanded.m
@@ -389,7 +389,7 @@ static NSMutableDictionary *colorNameCache = nil;
 	NSScanner *scanner = [NSScanner scannerWithString:stringToConvert];
 	if (![scanner scanString:@"{" intoString:NULL]) return nil;
 	const NSUInteger kMaxComponents = 4;
-	CGFloat c[kMaxComponents];
+	float c[kMaxComponents];
 	NSUInteger i = 0;
 	if (![scanner scanFloat:&c[i++]]) return nil;
 	while (1) {
